backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Video Processing
opencv-python-headless==4.9.0.80
pillow==10.2.0
ffmpeg-python==0.2.0

# GPS & Geospatial
geopy==2.4.1

# Storage
boto3==1.34.34
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
ruff==0.1.11
gunicorn==21.2.0
